Moved itemView out of AdminBookings
