+++ frontend/src/components/Admin/ItemView/ItemView.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from 'react';
 import { Link, useNavigate, useParams } from 'react-router-dom';
-import { useAppDispatch, useAppSelector } from '../../../../store/hooks';
-import { formatDate } from '../../../../utility/formatDate';
+import { useAppDispatch, useAppSelector } from '../../../store/hooks';
+import { formatDate } from '../../../utility/formatDate';
 import {
   deleteItem,
   fetchAllCategories,
@@ -9,7 +9,7 @@ import {
   selectAllCategories,
   selectItemById,
   updateItem,
-} from '../../../../slices/itemsSlice';
+} from '../../../slices/itemsSlice';
 import {
   Box,
   Button,
@@ -24,8 +24,8 @@ import {
 import { ImPencil2 } from 'react-icons/im';
 import { CiTrash } from 'react-icons/ci';
 import { ArrowBack } from '@mui/icons-material';
-import Spinner from '../../../Spinner';
-import { useAuth } from '../../../../hooks/useAuth';
+import Spinner from '../../Spinner';
+import { useAuth } from '../../../hooks/useAuth';
 import Slider from "react-slick";
 import "slick-carousel/slick/slick.css";
 import "slick-carousel/slick/slick-theme.css";
